I think this is my final commit in here. Tried groups didn't workout as expected. Discovered a bunch of misalignmets between books architecture and how GPU rendering is performed. Wrote a simple parser to load light obj models. Refractions the way they are computed are not compatible with loaded triangle mesh. Mark some of math as host device to write the parser. I am so done with this project my eyes hurt.

+++ src/ray.cu
@@ -115,7 +115,7 @@ __device__ bool ray::intersects(const world& w, intersection_list& intersect_lis
 {
 	bool intersect = false;
 
-	for (int i = 0; i < w.tail_element_index; i++)
+	for (size_t i = 0; i < w.tail_element_index; i++)
 	{
 		if (intersects(w.list[i], intersect_list))
 			intersect = true;
